fix: SQLite serial primary keys must map to INTEGER, not BIGINT
Dialect.Sqlite.saType mapped Serial8 to sa.BigInteger() (→ BIGINT).
SQLite only autoincrements the INTEGER PRIMARY KEY rowid alias; a
BIGINT PK is not that alias, so `autoincrement=True` was a silent
no-op and an app omitting the PK on insert would fail. The Alembic
DDL was still valid (the up/down/up CI round-trip is green — this is
a runtime-insert behavior), which is why it was latent.

Serial8 → sa.Integer() for SQLite: the rowid alias is itself 64-bit,
so the mapping is lossless, and this now matches the raw-SQL renderer
(Dialect.Sqlite.serialColumnDef already emits INTEGER PRIMARY KEY
AUTOINCREMENT for go-chi / ts-express). Postgres/MySQL saType
unchanged; AlembicMigrationTest's Postgres path still asserts
sa.BigInteger().

- DialectProfileEmitTest sqlite: assert the migration uses
  sa.Integer() for the autoincrement PK and contains no
  sa.BigInteger() (deterministic regression guard).
- Regenerated python-fastapi-sqlite golden (only 001_initial_schema
  id column changes).
- sqlite.mdx: replaced the (now-fixed) "Known limitation" note with
  a "Serial primary keys" delta entry documenting the correct
  behavior.

Tests: EmitPythonTest 3/3, DialectProfileEmitTest 9/9,
AlembicMigrationTest 8/8, EmitTest 21/21, EmitGo/EmitTs green.

+++ fixtures/golden/codegen/python/fastapi/sqlite/url_shortener/alembic/versions/001_initial_schema.py
@@ -21,7 +21,7 @@ def upgrade() -> None:
     op.create_table(
         "url_mappings",
 
-        sa.Column("id", sa.BigInteger(), primary_key=True, autoincrement=True, nullable=False),
+        sa.Column("id", sa.Integer(), primary_key=True, autoincrement=True, nullable=False),
 
         sa.Column("code", sa.Text(), nullable=False),
